fix(ext/crypto): deriveBits length arg + reason for clippy allow
Two CI fixes:

1. `clippy::allow_attributes_without_reason` flagged the
   `#[allow(clippy::too_many_arguments)]` on `derive_bits_sync`;
   add the `reason = ...` field.

2. Unit tests (`testHkdfDeriveBits`, `testEcdhDeriveBitsWith*`,
   `testDeriveKey`, `x25519SharedSecret`, `chaCha20Poly1305DeriveKey`)
   were failing with `Invalid length` because the
   `length: Option<u32>` parameter on the `deriveBits` /
   `__deriveBitsInternal` cppgc methods was binding to `None` even
   when JS passed a number. Bind it as `Option<f64>` (the native JS
   Number shape) and convert to `Option<u32>` inside
   `subtle_derive_bits::run` before the per-algorithm dispatch.

+++ ext/crypto/subtle_crypto.rs
@@ -171,7 +171,7 @@ impl SubtleCrypto {
     &self,
     #[webidl] algorithm: SubtleDeriveBitsParams,
     #[webidl] base_key: SubtleKey,
-    length: Option<u32>,
+    length: Option<f64>,
   ) -> Result<Vec<u8>, CryptoError> {
     if !base_key.has_usage("deriveBits") {
       return Err(CryptoError::Other(deno_error::JsErrorBox::new(
@@ -194,7 +194,7 @@ impl SubtleCrypto {
     &self,
     #[webidl] algorithm: SubtleDeriveBitsParams,
     #[webidl] base_key: SubtleKey,
-    length: Option<u32>,
+    length: Option<f64>,
   ) -> Result<Vec<u8>, CryptoError> {
     spawn_blocking(move || run_derive_bits(algorithm, base_key, length)).await?
   }
